refactor txscript runtime metering into versioned resource meters
