+++ crypto/txscript/src/lib.rs
@@ -14,7 +14,7 @@ pub mod standard;
 pub mod wasm;
 pub mod zk_precompiles;
 
-pub mod runtime_sig_op_counter;
+pub mod runtime_resource_meter;
 
 use std::io::Write;
 
@@ -45,7 +45,7 @@ pub mod prelude {
     pub use super::standard::*;
 }
 pub use crate::data_stack::{deserialize_i64, serialize_i64};
-use crate::runtime_sig_op_counter::RuntimeSigOpCounter;
+use crate::runtime_resource_meter::RuntimeResourceMeter;
 pub use crate::seq_commit_accessor::SeqCommitAccessor;
 pub use standard::*;
 
@@ -133,13 +133,8 @@ pub struct TxScriptEngine<'a, T: VerifiableTransaction, Reused: SigHashReusedVal
     cond_stack: Vec<OpCond>, // Following if stacks, and whether it is running
 
     num_ops: i32,
-    used_sig_ops: u16,
-    used_script_units: ScriptUnits,
-    accounted_pushed_bytes: u64,
-    runtime_sig_op_counter: RuntimeSigOpCounter,
+    runtime_resource_meter: RuntimeResourceMeter,
     opcode_execution_log_buffer: Option<&'a mut dyn Write>,
-    allowed_script_units: ScriptUnits,
-    remaining_script_units: ScriptUnits,
     flags: EngineFlags,
 }
 
@@ -369,6 +364,11 @@ pub fn is_unspendable<T: VerifiableTransaction, Reused: SigHashReusedValues>(scr
 
 impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'a, T, Reused> {
     pub fn new(ctx: EngineContext<'a, Reused>, flags: EngineFlags) -> Self {
+        let runtime_resource_meter = if flags.covenants_enabled {
+            RuntimeResourceMeter::new_script_units(flags.sigop_script_units, ScriptUnits(u64::MAX))
+        } else {
+            RuntimeResourceMeter::new_sigops(u8::MAX)
+        };
         Self {
             dstack: Self::new_stack(flags),
             astack: Self::new_stack(flags),
@@ -376,12 +376,7 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
             ctx,
             cond_stack: vec![],
             num_ops: 0,
-            used_sig_ops: 0,
-            used_script_units: ScriptUnits(0),
-            accounted_pushed_bytes: 0,
-            runtime_sig_op_counter: RuntimeSigOpCounter::new(u8::MAX),
-            allowed_script_units: ScriptUnits(u64::MAX),
-            remaining_script_units: ScriptUnits(u64::MAX),
+            runtime_resource_meter,
             flags,
             opcode_execution_log_buffer: None,
         }
@@ -393,12 +388,12 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
 
     /// Returns the number of signature operations used in script execution.
     pub fn used_sig_ops(&self) -> u16 {
-        self.used_sig_ops
+        self.runtime_resource_meter.used_sig_ops()
     }
 
     /// Returns the total script units consumed so far.
     pub fn used_script_units(&self) -> ScriptUnits {
-        self.used_script_units
+        self.runtime_resource_meter.used_script_units()
     }
 
     /// Returns the total bytes pushed into the data and alt stacks so far.
@@ -452,7 +447,11 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
         flags: EngineFlags,
         allowed_script_units: ScriptUnits,
     ) -> Self {
-        let allowed_script_units = if flags.covenants_enabled { allowed_script_units } else { ScriptUnits(u64::MAX) };
+        let runtime_resource_meter = if flags.covenants_enabled {
+            RuntimeResourceMeter::new_script_units(flags.sigop_script_units, allowed_script_units)
+        } else {
+            RuntimeResourceMeter::new_sigops(input.mass.sig_op_count().unwrap_or(0))
+        };
         let script_public_key = utxo_entry.script_public_key.script();
         // The script_public_key in P2SH is just validating the hash on the OpMultiSig script
         // the user provides
@@ -465,12 +464,7 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
             ctx,
             cond_stack: Default::default(),
             num_ops: 0,
-            used_sig_ops: 0,
-            used_script_units: ScriptUnits(0),
-            accounted_pushed_bytes: 0,
-            runtime_sig_op_counter: RuntimeSigOpCounter::new(input.mass.sig_op_count().unwrap_or(0)),
-            allowed_script_units,
-            remaining_script_units: allowed_script_units,
+            runtime_resource_meter,
             opcode_execution_log_buffer: None,
             flags,
         }
@@ -492,6 +486,11 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
         allowed_script_units: ScriptUnits,
         flags: EngineFlags,
     ) -> Self {
+        let runtime_resource_meter = if flags.covenants_enabled {
+            RuntimeResourceMeter::new_script_units(flags.sigop_script_units, allowed_script_units)
+        } else {
+            RuntimeResourceMeter::new_sigops(u8::MAX)
+        };
         Self {
             dstack: Self::new_stack(flags),
             astack: Self::new_stack(flags),
@@ -499,40 +498,21 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
             ctx: EngineCtx::new(sig_cache).with_reused(reused_values),
             cond_stack: Default::default(),
             num_ops: 0,
-            used_sig_ops: 0,
-            used_script_units: ScriptUnits(0),
-            accounted_pushed_bytes: 0,
-            runtime_sig_op_counter: RuntimeSigOpCounter::new(u8::MAX),
-            allowed_script_units,
-            remaining_script_units: allowed_script_units,
+            runtime_resource_meter,
             opcode_execution_log_buffer: None,
             flags,
         }
     }
 
     fn consume_script_units(&mut self, units: ScriptUnits) -> Result<(), TxScriptError> {
-        match self.remaining_script_units.checked_sub(units) {
-            Some(new_remaining) => {
-                self.used_script_units = self.used_script_units.saturating_add(units);
-                self.remaining_script_units = new_remaining;
-                Ok(())
-            }
-            None => {
-                let overflow = units - self.remaining_script_units;
-                let used_units = self.allowed_script_units + overflow;
-                Err(TxScriptError::ExceededScriptUnitsLimit { used_units: used_units.0, allowed_units: self.allowed_script_units.0 })
-            }
+        if let RuntimeResourceMeter::ScriptUnits(meter) = &mut self.runtime_resource_meter {
+            meter.consume_script_units(units)?;
         }
+        Ok(())
     }
 
     fn consume_sig_op_cost(&mut self, count: u16) -> Result<(), TxScriptError> {
-        if self.flags.covenants_enabled {
-            self.consume_script_units(ScriptUnits((count as u64).saturating_mul(self.flags.sigop_script_units.0)))?;
-        } else {
-            self.runtime_sig_op_counter.consume_sig_ops(count)?;
-        }
-        self.used_sig_ops = self.used_sig_ops.saturating_add(count);
-        Ok(())
+        self.runtime_resource_meter.consume_sig_op_cost(count)
     }
 
     #[inline]
@@ -559,11 +539,9 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
                 opcode.check_minimal_data_push()?;
             }
             opcode.execute(self)?;
-            if self.flags.covenants_enabled {
-                let total_pushed_bytes = self.total_pushed_bytes();
-                let pushed_bytes_delta = total_pushed_bytes.saturating_sub(self.accounted_pushed_bytes);
-                self.consume_script_units(pushed_bytes_delta.into())?;
-                self.accounted_pushed_bytes = total_pushed_bytes;
+            let total_pushed_bytes = self.total_pushed_bytes();
+            if let RuntimeResourceMeter::ScriptUnits(meter) = &mut self.runtime_resource_meter {
+                meter.charge_newly_pushed_bytes(total_pushed_bytes)?;
             }
             Ok(())
         } else {
